Allow to read-out huge memory allocations.
When running the example in the top directory PROLEAD is crashing by throwing a std::out_of_range.
While clearly something is wrong with the amount of allocated data (21475178120 kB on my WSL2 based system)
it can be prevented from crashing by reading out the allocated memory in an unsigned 64 bits value.

This is what his patch does.